fix(chat): address CodeRabbit review feedback
